refactor: restructure chat client modules and consolidate conversation handling
- Moved conversation-related classes into a dedicated `conversation` module for better organization.
- Introduced `RagbitsChatClient` and `AsyncRagbitsChatClient` to manage synchronous and asynchronous chat interactions.
- Removed the old `conversation.py` file, streamlining the codebase and enhancing maintainability.
- Updated imports in the chat client to reflect the new structure, ensuring clarity and consistency across modules.
