+++ packages/ragbits-chat/src/ragbits/chat/clients/__init__.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-from .async_client import AsyncRagbitsChatClient
+from .client import AsyncRagbitsChatClient, RagbitsChatClient
 from .conversation import AsyncConversation, Conversation
-from .sync_client import RagbitsChatClient
 
 __all__ = [
     "AsyncConversation",
