+++ packages/ragbits-chat/src/ragbits/chat/clients/client/__init__.py
@@ -1,0 +1,7 @@
+from .async_client import AsyncRagbitsChatClient
+from .sync_client import RagbitsChatClient
+
+__all__ = [
+    "AsyncRagbitsChatClient",
+    "RagbitsChatClient",
+]
