+++ packages/ragbits-chat/src/ragbits/chat/clients/conversation/__init__.py
@@ -1,0 +1,7 @@
+from .async_conversation import AsyncConversation
+from .sync_conversation import Conversation
+
+__all__ = [
+    "AsyncConversation",
+    "Conversation",
+]
